ng2: move app.ng2module to new file
