+++ generators/client-2/templates/src/main/webapp/app/_app.module.ts
@@ -5,14 +5,6 @@ import './entities/entity.module';
 
 import { upgradeAdapter } from "./upgrade_adapter";
 
-import { XSRFStrategy, CookieXSRFStrategy } from '@angular/http';
-import { FormsModule } from '@angular/forms';
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-
-import { <%=angularAppName%>CommonModule } from './components/common.module';
-import { <%=angularAppName%>AdminModule } from './admin/admin.module';
-
 import { StateHandler } from './blocks/handlers/state.handler';<% if (enableTranslation) { %>
 import { TranslationHandler } from './blocks/handlers/translation.handler';
 
@@ -38,8 +30,6 @@ import { NotificationInterceptor } from './blocks/interceptor/notification.inter
 import { AppStateConfig } from './app.state';
 import { HomeStateConfig } from './home/home.state';
 import { ErrorStateConfig } from './layouts/error/error.state';
-
-import {FindLanguageFromKeyPipe} from "./components/language/language.pipe";
 
 angular
     .module('<%=angularAppName%>.app', [
@@ -90,12 +80,3 @@ function run(StateHandler<% if (enableTranslation) { %>, TranslationHandler<% } 
     StateHandler.initialize();<% if (enableTranslation) { %>
     TranslationHandler.initialize();<% } %>
 }
-
-@NgModule({
-    imports: [BrowserModule, FormsModule, <%=angularAppName%>CommonModule, <%=angularAppName%>AdminModule],
-    providers: [{
-        provide: XSRFStrategy, useValue:  new CookieXSRFStrategy('CSRF-TOKEN', 'X-CSRF-TOKEN')
-    }],
-    declarations: [FindLanguageFromKeyPipe]
-})
-export class <%=angularAppName%>AppModule {}
